requirements.txt
fastapi
uvicorn
mysql-connector-python
schedule
requests
python-dotenv